Resolves issue #64
When compass returns a compile error it seems to do so in standardOutput.

The issue is that SassyStudio only outputs the errorOutput OR the standardOutput. If the errorOutput.Length is greater than zero then the standardOutput is never displayed to the user. It seems that the errorOutput.Length is always greater than zero (because of newline characters).